change janpai.js to janpai.ts
