+++ src/main/ts/jan/hand-util.js
@@ -15,7 +15,7 @@ limitations under the License.
  */
 import CompleteInfo from './complete-info';
 import CompletePattern from './complete-pattern';
-import Janpai from './janpai';
+import {Janpai} from './janpai';
 import {JANPAI_ID} from './janpai-id';
 import JanUtil from './jan-util';
 import KnittedStraight from './knitted-straight';
