+++ src/main/ts/jan/jan-util.js
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
  */
-import Janpai from './janpai';
+import {Janpai} from './janpai';
 import {JANPAI_ID} from './janpai-id';
 import Mentsu from './mentsu';
 import {WIND} from './wind';
