+++ src/test/js/jan/hand-util-test.js
@@ -19,7 +19,7 @@ import CompleteInfo from '../../../main/js/jan/complete-info';
 import CompletePattern from '../../../main/js/jan/complete-pattern';
 import Hand from '../../../main/js/jan/hand';
 import HandUtil from '../../../main/js/jan/hand-util';
-import Janpai from '../../../main/js/jan/janpai';
+import {Janpai} from '../../../main/js/jan/janpai';
 import {JANPAI_ID} from '../../../main/js/jan/janpai-id';
 import Mentsu from '../../../main/js/jan/mentsu';
