+++ src/test/js/jan/mcr-util-test.js
@@ -19,7 +19,7 @@ import CompleteInfo from '../../../main/js/jan/complete-info';
 import CompletePattern from '../../../main/js/jan/complete-pattern';
 import {COMPLETE_TYPE} from '../../../main/js/jan/complete-type';
 import Hand from '../../../main/js/jan/hand';
-import Janpai from '../../../main/js/jan/janpai';
+import {Janpai} from '../../../main/js/jan/janpai';
 import {JANPAI_ID} from '../../../main/js/jan/janpai-id';
 import McrUtil from '../../../main/js/jan/mcr-util';
 import {MCR_YAKU} from '../../../main/js/jan/mcr-yaku';
